backend/requirements.txt
fastapi
"uvicorn[standard]"
python-multipart
pdf2image
pillow
opencv-python
numpy
